iot_water_monitoring_system/pyproject.toml
[tool.poetry]
name = "iot-water-monitoring-system"
version = "0.1.0"
description = ""
authors = ["Dhruv Patel <pateldhruv130401@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
paho-mqtt = "<2.0"
pandas = "^2.2.3"
scikit-learn = "^1.5.2"
plotly = "^5.24.1"
channels = "^4.2.0"
uvicorn = "^0.32.0"
wsproto = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
